+++ examples/semantic_test_range_product.rs
@@ -3092,8 +3092,8 @@ pub fn range_product(start: i32, end: i32) -> i32 {
     result = 1;
     let mut i: i32 = start.clone();
     while i <= end {
-        result = (result).py_mul(i);
-        i = (i).py_add(1);
+        result = ((result).py_mul(i)) as i32;
+        i = ((i).py_add(1i32)) as i32;
     }
     result
 }
