+++ examples/semantic_test_strictly_decreasing.rs
@@ -3112,7 +3112,7 @@ pub fn strictly_decreasing(nums: &Vec<i32>) -> Result<bool, Box<dyn std::error::
             .expect("IndexError: list index out of range")
             >= {
                 let base = &nums;
-                let idx: i32 = (i).py_sub(1);
+                let idx: i32 = (i).py_sub(1i32);
                 let actual_idx = if idx < 0 {
                     base.len().saturating_sub(idx.abs() as usize)
                 } else {
@@ -3125,7 +3125,7 @@ pub fn strictly_decreasing(nums: &Vec<i32>) -> Result<bool, Box<dyn std::error::
         {
             return Ok(false);
         }
-        i = (i).py_add(1);
+        i = ((i).py_add(1i32)) as i32;
     }
     Ok(true)
 }
